+++ src/io/GLGPU_IO_Helper.h
@@ -7,23 +7,23 @@
 bool GLGPU_IO_Helper_ReadBDAT(
     const std::string& filename, 
     GLHeader &hdr,
-    double **rho, double **phi, double **re, double **im, 
+    float **rho, float **phi, float **re, float **im, 
     bool header_only=false);
 
 bool GLGPU_IO_Helper_ReadLegacy(
     const std::string& filename, 
     GLHeader &hdr, 
-    double **rho, double **phi, double **re, double **im, 
+    float **rho, float **phi, float **re, float **im, 
     bool header_only=false);
 
 bool GLGPU_IO_Helper_ReadNetCDF(
     const std::string& filename, 
     GLHeader &hdr, 
-    double **psi); 
+    float **psi); 
 
 bool GLGPU_IO_Helper_WriteNetCDF(
     const std::string& filename, 
     GLHeader &hdr, 
-    double **psi); 
+    float **psi); 
 
 #endif
